requirements.txt
pydantic-settings
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
python-multipart
